requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
gast==0.3.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.3.0
inflection==0.3.1
ipykernel==5.1.3
ipython==7.11.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==5.0.0
nbconvert==5.6.1
nbformat==4.4.0
ndg-httpsclient==0.5.1
notebook==6.0.2
numpy==1.18.0
opencv-python==4.1.2.30
opencv-utils==0.0.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pkg-resources==0.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.2
psycopg2==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.5.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.6
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
Quandl==3.4.8
requests==2.22.0
scikit-learn==0.22
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
SQLAlchemy==1.3.12
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
